+++ Runtime/_Experimental/KeyFluxT.cs
@@ -30,10 +30,12 @@ namespace Kingdox.UniFlux.Experimental
         public sealed override void Store(in Action callback, in bool condition) => Core.Flux.Store(KeyT, callback, condition);
         public sealed override void Dispatch() => Core.Flux.Dispatch(KeyT);
         public sealed override void Store<T2>(in Action<T2> callback, in bool condition) => Core.Flux.Store<T,T2>(KeyT, in callback, in condition);
-        public sealed override void Dispatch<T2>(in T2 @param) => Core.Flux.Dispatch(in @param);
+        public sealed override void Dispatch<T2>(in T2 @param) => Core.Flux.Dispatch(KeyT, in @param);
         public sealed override void Store<T2>(in Func<T2> callback, in bool condition) => Core.Flux.Store<T,T2>(KeyT, in callback, in condition);
         public sealed override T2 Dispatch<T2>() => Core.Flux.Dispatch<T, T2>(KeyT);
         public sealed override void Store<T2, T3>(in Func<T2, T3> callback, in bool condition) => Core.Flux.Store<T,T2,T3>(KeyT, in callback, in condition);
         public sealed override T3 Dispatch<T2, T3>(in T2 @param) => Core.Flux.Dispatch<T,T2,T3>(KeyT, in @param);
+        public sealed override void StoreState<T2>(in Action<T2> callback, in bool condition) => Core.Flux.StoreState<T,T2>(KeyT, in callback, in condition);
+        public sealed override void DispatchState<T2>(in T2 @param) => Core.Flux.DispatchState(KeyT, in @param);
     }
 }